openconnect: fix parallel compilation failure
This patch solves the following error which appears now and then because of
parallel building.

  ../git/main.c:78:21: fatal error: version.c: No such file or directory

version.c is generated at build time. And it's needed by some objects. So
it needs to be generated first to avoid compilation failure.

This patch fix the problem by generating version.c first before building others.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>